Port: Part 1. gregtech.api, gregtech.common.covers, gregtech.common.render

+++ src/main/java/gregtech/common/covers/GT_Cover_RedstoneReceiverExternal.java
@@ -6,7 +6,7 @@ import gregtech.api.interfaces.tileentity.ICoverable;
 public class GT_Cover_RedstoneReceiverExternal
         extends GT_Cover_RedstoneWirelessBase {
     public int doCoverThings(byte aSide, byte aInputRedstone, int aCoverID, int aCoverVariable, ICoverable aTileEntity, long aTimer) {
-        aTileEntity.setOutputRedstoneSignal(aSide, GregTech_API.sWirelessRedstone.get(Integer.valueOf(aCoverVariable)) == null ? 0 : ((Byte) GregTech_API.sWirelessRedstone.get(Integer.valueOf(aCoverVariable))).byteValue());
+        aTileEntity.setOutputRedstoneSignal(aSide, GregTech_API.sWirelessRedstone.get(aCoverVariable) == null ? 0 : GregTech_API.sWirelessRedstone.get(aCoverVariable));
         return aCoverVariable;
     }
