+++ src/main/java/gregtech/common/covers/GT_Cover_RedstoneTransmitterExternal.java
@@ -3,10 +3,10 @@ package gregtech.common.covers;
 import gregtech.api.GregTech_API;
 import gregtech.api.interfaces.tileentity.ICoverable;
 
-public class GT_Cover_RedstoneTransmitterExternal
-        extends GT_Cover_RedstoneWirelessBase {
+public class GT_Cover_RedstoneTransmitterExternal extends GT_Cover_RedstoneWirelessBase {
+
     public int doCoverThings(byte aSide, byte aInputRedstone, int aCoverID, int aCoverVariable, ICoverable aTileEntity, long aTimer) {
-        GregTech_API.sWirelessRedstone.put(Integer.valueOf(aCoverVariable), Byte.valueOf(aInputRedstone));
+        GregTech_API.sWirelessRedstone.put(aCoverVariable, aInputRedstone);
         return aCoverVariable;
     }
